requirements.txt
-i https://pypi.org/simple
arrow==0.15.6
asgiref==3.2.7
awesome-slugify==1.6.5
beautifulsoup4==4.9.1
blessed==1.17.6
cachetools==4.1.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cryptography==2.8
cssselect==1.1.0
cssutils==1.0.2
decorator==4.4.2
django-picklefield==2.1.1
django-q-sentry==0.1.1
django-q[sentry]==1.2.1
django-redis==4.11.0
django-simple-history==2.10.0
django-webpack-loader==0.7.0
django==3.0.4
feedfinder2==0.0.4
feedparser==5.2.1
future==0.18.2
gdal==2.4.0
gunicorn==20.0.4
h11==0.9.0
httptools==0.1.1 ; sys_platform != 'win32' and sys_platform != 'cygwin' and platform_python_implementation != 'PyPy'
idna==2.9
jieba3k==0.35.1
lxml==4.5.1
mistune==2.0.0a4
newspaper3k==0.2.8
nltk==3.4.5
patreon==0.5.0
pillow==7.1.2
premailer==3.6.1
psycopg2-binary==2.8.4
pycparser==2.20
pyjwt==1.7.1
python-dateutil==2.8.1
python-dotenv==0.13.0
python-telegram-bot==12.5.1
pytz==2020.1
pyyaml==5.3.1
redis==3.5.2
regex==2020.5.14
requests-file==1.5.1
requests==2.23.0
sentry-sdk==0.14.3
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
tinysegmenter==0.3
tldextract==2.2.2
tornado==6.0.4
typus==0.2.2
unidecode==0.4.21
urllib3==1.25.9
uvicorn==0.11.5
uvloop==0.14.0 ; sys_platform != 'win32' and sys_platform != 'cygwin' and platform_python_implementation != 'PyPy'
wcwidth==0.1.9
websockets==8.1
